改一波

+++ src/views/home/index.vue
@@ -183,9 +183,20 @@ export default {
       })
     },
     handleDelete(id) {
-      deleteInfo(id).then(res => {
-        this.$message.success('删除成功')
-        this.initData()
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteInfo(id).then(res => {
+          this.$message.success('删除成功')
+          this.initData()
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     },
     addFormShow() {
